+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   and methods. This should make it easier to disambiguate between error types.
 - `EncodeDbn::encode_record` and `EncodeDbn::record_record_ref` no longer treat a
   `BrokenPipe` error differently
+- Added `pretty_` Python attributes for DBN price fields
 
 ### Breaking changes
 - All fallible operations now return a `dbn::Error` instead of an `anyhow::Error`